// infrastructure/mappers/movieMapper.ts
import { Movie } from "../interfaces/Movie";
import { Result } from "../interfaces/movieDbResponse";
import { MovieDetail } from "../interfaces/MovieDetail";
import { MovieDetailResponse } from "../interfaces/MovieDetailResponse";

export class MovieMapper {
  static fromMovieDBtoMovie(movie: Result): Movie {
    return {
      id: movie.id,
      title: movie.title,
      description: movie.overview,
      releaseDate: new Date(movie.release_date),
      rating: movie.vote_average,
      poster: `https://image.tmdb.org/t/p/w500${movie.poster_path}`,
      backdrop: `https://image.tmdb.org/t/p/w500${movie.backdrop_path}`,
    };
  }
  static fromMovieDetailResponse(movie: MovieDetailResponse): MovieDetail {
    return {
      id: movie.id,
      title: movie.title,
      description: movie.overview,
      releaseDate: new Date(movie.release_date),
      rating: movie.vote_average,
      poster: `https://image.tmdb.org/t/p/w500${movie.poster_path}`,
      backdrop: `https://image.tmdb.org/t/p/w500${movie.backdrop_path}`,
      budget: movie.budget,
      duration: movie.runtime,
      genres: movie.genres.map((g) => g.name),
      originalTitle: movie.original_title,
      productionCompanies: movie.production_companies.map((p) => p.name),
    };
  }
}
